Refactor : updated the tests so that LIC,PUM,FUV and DECIDE tests are separated and are in different folders #46
